fix: remove duplicated types and consolidate into one file
