--- types/api/gamma/currency.ts
@@ -1,50 +1,0 @@
-export type TokenMetaData = {
-  image_url: string;
-  image_type: string;
-  image_protocol: string;
-  asset_url: string;
-  asset_type: string;
-  asset_protocol: string;
-  asset_id: string;
-  name: string;
-  contract_id: string;
-};
-
-export type Attribute = {
-  value: string;
-  trait_type: string;
-};
-
-export type NftData = {
-  asset_id: string;
-  collection_contract_id: string;
-  token_id: number;
-  fully_qualified_token_id: string;
-  token_metadata: TokenMetaData;
-  nft_token_attributes: Attribute[];
-  rarity_rank: string;
-  collection_count: number;
-  rarity_score: string;
-};
-
-export type NftDetailResponse = {
-  result: boolean;
-  data: NftData;
-  isCompliant?: boolean;
-};
-
-export interface NftCollectionData {
-  collection: {
-    id: string;
-    contractId: string;
-    name: string;
-    description: string;
-    image: string;
-    verified: false;
-    link: string;
-    floorItem: {
-      nftId: string;
-      price: string;
-    };
-  };
-}
